Spread: added basic spread test definition
This commit enables the use of spread to test pebble.

For now the only implemented spread backend is an LXD container.
The existing pebble unit tests are run in a spread test suite, as well
as two basic integration tests.
The pebble service is started and stopped for each individual
integration test.

To try it yourself, run spread from the root of this repo.