HACK: make websockets work again
Websockets used to work on the demo branch, but a minor websocket
workaround failed to make it to this temporary development enabler branch.

Add a custom TCP dialer to allow the current limited implementation
of websockets to work with remote TCP connections. This provides KernOS
environments with very useful debug abilities, which are critical to
have at this stage of KernOS development. The example below assumes
suitable debug tools are installed on the KernOS appliance:

PEBBLE_SOCKET=<IP>:<PORT> ./kernos exec /bin/sh